+++ src/RedirectUriValidators/RedirectUriValidator.php
@@ -9,6 +9,9 @@
 
 namespace League\OAuth2\Server\RedirectUriValidators;
 
+use League\Uri\Exceptions\SyntaxError;
+use League\Uri\Uri;
+
 class RedirectUriValidator implements RedirectUriValidatorInterface
 {
     /**
@@ -59,10 +62,14 @@ class RedirectUriValidator implements RedirectUriValidatorInterface
      */
     private function isLoopbackUri($redirectUri)
     {
-        $parsedUrl = \parse_url($redirectUri);
+        try {
+            $uri = Uri::createFromString($redirectUri);
+        } catch (SyntaxError $e) {
+            return false;
+        }
 
-        return $parsedUrl['scheme'] === 'http'
-            && (\in_array($parsedUrl['host'], ['127.0.0.1', '[::1]'], true));
+        return $uri->getScheme() === 'http'
+            && (\in_array($uri->getHost(), ['127.0.0.1', '[::1]'], true));
     }
 
     /**
@@ -106,9 +113,8 @@ class RedirectUriValidator implements RedirectUriValidatorInterface
      */
     private function parseUrlAndRemovePort($url)
     {
-        $parsedUrl = \parse_url($url);
-        unset($parsedUrl['port']);
+        $uri = Uri::createFromString($url);
 
-        return $parsedUrl;
+        return (string) $uri->withPort(null);
     }
 }